requirements.txt
websocket-client
requests
yt-dlp